Refactor comments & namespaces
